Изменен тип переменных height, weight. Убрана пустая строка в конце кода
Добавлен пробел между скобками ){

+++ Lesson_2/Person.java
@@ -1,8 +1,8 @@
 public class Person {
     String sex = "мужской";
     String name = "Иван";
-    int height = 180;
-    int weight = 78;
+    float height = 180;
+    float weight = 78;
     int age = 32;
 
     public void go() {
@@ -13,7 +13,7 @@ public class Person {
         System.out.println(name + " сидит");
     }
 
-    public void run(){
+    public void run() {
         System.out.println(name + " бежит");
     }
 
@@ -24,5 +24,4 @@ public class Person {
     public void learnJava() {
         System.out.println(name + " учит Java");
     }
-
 }
